Abstract away logger
